requirements.txt
appdirs==1.4.3
docopt==0.6.2
future==0.16.0
jedi==0.10.2
oauthlib==2.0.2
packaging==16.8
prompt-toolkit==1.0.14
ptpython==0.39
Pygments==2.2.0
pyparsing==2.2.0
python-twitter==3.3
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
wcwidth==0.1.7
